Rename `partial2` module to `partial`
