+++ crates/fj-kernel/src/algorithms/intersect/ray_face.rs
@@ -155,7 +155,7 @@ mod tests {
         builder::FaceBuilder,
         insert::Insert,
         iter::ObjectIters,
-        partial2::{PartialFace, PartialObject},
+        partial::{PartialFace, PartialObject},
         services::Services,
     };
