+++ crates/fj-kernel/src/lib.rs
@@ -98,7 +98,7 @@ pub mod geometry;
 pub mod insert;
 pub mod iter;
 pub mod objects;
-pub mod partial2;
+pub mod partial;
 pub mod services;
 pub mod storage;
 pub mod validate;
