+++ crates/fj-kernel/src/partial/mod.rs
