+++ crates/fj-kernel/src/partial/traits.rs
